# contracts/openapi/policy.yaml
openapi: 3.1.0
info:
  title: SETI - Policy Job
  description: |
    Owns the primary SETI database. The Policy Job is the administrative
    core of the SETI constellation — it holds the application registry,
    test schedules, failure response policies, and clearance definitions.

    Admin access only. Never externally exposed. The only authorized
    callers are the Gateway (for admin operations) and internal SETI
    Jobs that need to read policy during test execution.

    The Policy Job owns four distinct concerns:

    1. APPLICATION REGISTRY — the list of TCA applications SETI monitors.
       Each entry holds the repository contracts path, gateway endpoint,
       observability stream connection, and environment configuration.

    2. SCHEDULES — when Contract Tests and Plot Tests run per application.
       Configurable per application, per test tier, per environment.
       Default: Contract Tests every 15 minutes, Plot Tests daily plus
       on deployment trigger.

    3. FAILURE RESPONSE POLICY — what SETI does when a test fails.
       Keyed by application, Job name, test tier, environment, and
       severity. Drives the Interactions Job routing decisions.

    4. CONTRACT VERSION TRACKING — the last known contract version per
       Job per application. Used by the Contract Test Job to detect
       version changes and flag Plot Store for regeneration.

    Internal service — mTLS only, not externally exposed.
  version: 0.1.0
...
servers:
...
paths:
...
  /applications:
    post:
      tags: [applications]
      summary: Register a new application for SETI monitoring
      description: |
        Registers a TCA application in the SETI monitoring registry.
        On registration, SETI gains read access to the application's
        contracts folder and begins scheduling tests per the provided
        schedule configuration.

        Registration does not immediately trigger a test run. The first
        scheduled run occurs at the next schedule interval.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationRegistration'
      responses:
        '201':
          description: Application registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: Invalid registration payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Application with this ID already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags: [applications]
      summary: List all registered applications
      parameters:
        - name: environment
          in: query
          required: false
          schema:
            type: string
            enum: [production, staging, development]
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, paused, error]
      responses:
        '200':
          description: Application list returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationList'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    ApplicationRegistration:
      type: object
      required:
        - application_id
        - display_name
        - repository_url
        - contracts_path
        - environments
      properties:
        application_id:
          type: string
          description: >
            Unique identifier for this application across all environments.
            Convention: app-{name}-{env} e.g. app-tca-vox-prod
          example: "app-tca-vox-prod"
        display_name:
          type: string
          description: Human-readable application name
          example: "TCA Vox (Production)"
        repository_url:
          type: string
          description: >
            Repository URL. SETI reads from the contracts folder only.
            Read access must be granted before registration.
          example: "https://github.com/Pelentan/tca-vox"
        contracts_path:
          type: string
          description: Path to the contracts folder within the repository
          example: "contracts/openapi"
        environments:
          type: array
          items:
            $ref: '#/components/schemas/EnvironmentConfig'
          minItems: 1
          description: One entry per environment being monitored
        ac_endpoint:
          type: string
          description: >
            HTTPS endpoint for this constellation's Augur Canis instance.
            Set when a Sec Wr4ngler registers the constellation. SETI uses
            this to initiate federation registration with the external AC.
            Required for health feed subscription. Optional for seti-self
            which uses the internal AC directly.
        federation_status:
          type: string
          enum: [pending, registered, degraded, unregistered]
          description: >
            pending — ac_endpoint provided, federation registration not yet
            attempted.
            registered — SETI has a valid session certificate from the AC,
            health feed active.
            degraded — feed silent or signature verification failing, SETI
            is attempting re-registration.
            unregistered — no ac_endpoint, or registration failed permanently.
        federation_registered_at:
          type: string
          format: date-time
          description: When the current session certificate was issued.
        session_cert_id:
          type: string
          description: >
            Short identifier of the session certificate currently in use.
            SETI uses this to detect certificate rotation in incoming events.
        default_schedule:
          $ref: '#/components/schemas/TestScheduleConfig'
          description: >
            Schedule applied to all environments unless overridden.
            If omitted, SETI defaults apply: contract tests every
            15 minutes, plot tests daily at 02:00 UTC.
    EnvironmentConfig:
      type: object
      required:
        - environment
        - gateway_endpoint
        - observability_stream
      properties:
        environment:
          type: string
          enum: [production, staging, development]
        gateway_endpoint:
          type: string
          description: Gateway URL for this environment
          example: "https://vox.example.com"
        observability_stream:
          type: string
          description: Redis connection string for tca:events pub/sub
          example: "redis://vox-redis-prod:6379"
        schedule_override:
          $ref: '#/components/schemas/TestScheduleConfig'
          description: >
            Environment-specific schedule. Overrides default_schedule
            for this environment only.
    TestScheduleConfig:
      type: object
      properties:
        contract_test_interval_minutes:
          type: integer
          minimum: 5
          maximum: 1440
          description: >
            How often to run Contract Tests in minutes.
            Minimum 5 minutes. Default 15.
          example: 15
        plot_test_cron:
          type: string
          description: >
            Cron expression for Plot Test runs. Default: daily at 02:00 UTC.
          example: "0 2 * * *"
        plot_test_on_deployment:
          type: boolean
          description: >
            Whether to trigger a Plot Test run on deployment detection.
            Default true.
          example: true
        timeout_seconds:
          type: integer
          minimum: 30
          maximum: 300
          description: >
            Maximum time a Contract Test run may take before timing out.
            Default 60 seconds.
          example: 60
...
    Application:
      type: object
      required:
        - application_id
        - display_name
        - repository_url
        - contracts_path
        - status
        - registered_at
        - updated_at
        - environments
      properties:
        application_id:
          type: string
        display_name:
          type: string
        repository_url:
          type: string
        contracts_path:
          type: string
        status:
          type: string
          enum: [active, paused, error]
          description: >
            active — tests running on schedule.
            paused — tests suspended, application still registered.
            error — SETI cannot reach repository or gateway.
        pause_reason:
          type: string
          description: Present when status is paused
        registered_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        environments:
          type: array
          items:
            $ref: '#/components/schemas/EnvironmentConfig'
        default_schedule:
          $ref: '#/components/schemas/TestScheduleConfig'
    ApplicationList:
      type: object
      required:
        - applications
        - total
      properties:
        applications:
          type: array
          items:
            $ref: '#/components/schemas/Application'
        total:
          type: integer
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "APPLICATION_NOT_FOUND"
        message:
          type: string
          example: "Application 'app-tca-vox-prod' not found in registry"


  # ---------------------------------------------------------------------------
  # AVAILABLE APPLICATIONS (remote-apps.json source)
  # ---------------------------------------------------------------------------
...